add leverage token data util
